refactor: datatables component file structure and naming
