+++ src/lib/components/datatables/TableRowCandidates.svelte
